+++ package.json
@@ -11,6 +11,7 @@
     "node": ">=16.14"
   },
   "dependencies": {
+    "@faker-js/faker": "^8.3.1",
     "@fontsource/roboto-mono": "^4.5.10",
     "@fortawesome/fontawesome-svg-core": "^6.4.0",
     "@fortawesome/free-regular-svg-icons": "^6.4.0",
@@ -92,8 +93,7 @@
     "webpack": "^5.75.0",
     "winston": "^3.8.2",
     "winston-loki": "^6.0.6",
-    "zod": "^3.20.2",
-    "@faker-js/faker": "^8.3.1"
+    "zod": "^3.20.2"
   },
   "scripts": {
     "dev": "run-p start api",
@@ -209,7 +209,7 @@
     "http-terminator": "^3.2.0",
     "jest": "^29.5.0",
     "keycloak-request-token": "^0.1.0",
-    "msw": "^2.1.3",
+    "msw": "^2.1.4",
     "nodemon": "^2.0.20",
     "npm-run-all": "^4.1.5",
     "prop-types": "^15.8.1",
