+++ src/api/jest.polyfills.ts
@@ -28,3 +28,5 @@ Object.defineProperties(globalThis, {
   Request: { value: Request },
   Response: { value: Response },
 });
+
+export {};
