+++ src/web/components/SharingPermission/SearchAndAddParticipants.stories.tsx
@@ -1,7 +1,7 @@
 import { ComponentMeta, ComponentStory } from '@storybook/react';
 
 import { SharingSiteDTO } from '../../../api/helpers/siteConvertingHelpers';
-import { AvailableParticipantDTO } from '../../../api/routers/participantsRouter';
+import { AvailableParticipantDTO } from '../../../api/routers/participants/participantsRouter';
 import { TestAvailableSiteListProvider } from '../../services/site';
 import { SearchAndAddParticipants } from './SearchAndAddParticipants';
